// App/src/app/main/apps/admission/Models/DotTuyenSinh.model.ts
import { BacDaoTao } from "./BacDaoTao.model";
import { HeDaoTao } from "./HeDaoTao.model";
import { HinhThucXetTuyen } from "./HinhThucXetTuyen.model";
import { NganhDaoTao } from "./NganhDaoTao.model";
import { GiayTo } from "./GiayTo.model";

export class DotTuyenSinh {
	Id: number;
	Ten: string;
	HeDaoTaoId: number;
	HeDaoTao: HeDaoTao;
	CongThucDiemId: number;
	CongThucDiem: string;
	HinhThucXetTuyenId: number;
	HinhThucXetTuyen: HinhThucXetTuyen;
	NganhDaoTao_ToHopCotDiems: any;
	DanhSachNganhDaoTao: Array<NganhDaoTao>;
	BacDaoTaoId: number;
	BacDaoTao: BacDaoTao;
	KhoaHoc: string;
	LoaiDangKyId: number;
	NamTuyenSinh: number;
	ThuTuHienThi: number;
	KichHoatDK: boolean;
	KichHoatXetTuyet: boolean;
	NgayTao: Date;
	Ngaysua: Date;
	NguoiTao: string;
	NguoiSua: string;
	SoLuongNguyenVong: number;
	HanDangKy: Date;
	NoiDungThongBao: string;
	NoiDungEmail: string;
	CongThucKiemTraDieuKienNhapDiem: string;
	ChoPhepGuiMail: boolean;
	ChoPhepNhapDCThuongTruVaTamTru: boolean;
	KiemTraDieuKienNhapDiem: boolean;
	NgayBatDau: Date;
	MaTienTo: string;
	MaBacDaoTao: string;
	MaBatDau: string;
	PhuongThucXetTuyenId: number;
	PhuongThucXetTuyens: any[];
	GiayToYeuCaus: Array<GiayTo>;
	FileHuongDan: GiayTo;
	IsTaiFile: boolean;
	ChoPhepUploadHoSo: boolean;
	IsKhoa: boolean;
	NgayKhoa: Date;
	PhanQuyenChinhSuaHoSo: number;
	constructor(dotTuyenSinh?) {
		dotTuyenSinh = dotTuyenSinh || {};
		this.Id = dotTuyenSinh.Id || null;
		this.Ten = dotTuyenSinh.Ten || '';
		this.HeDaoTaoId = dotTuyenSinh.HeDaoTaoId || '';
		this.HeDaoTao = dotTuyenSinh.HeDaoTao || '';
		this.CongThucDiemId = dotTuyenSinh.CongThucDiemId || '';
		this.CongThucDiem = dotTuyenSinh.CongThucDiem || '';
		this.HinhThucXetTuyenId = dotTuyenSinh.HinhThucXetTuyenId || '';
		this.HinhThucXetTuyen = dotTuyenSinh.HinhThucXetTuyen || new HinhThucXetTuyen();
		this.BacDaoTao = dotTuyenSinh.BacDaoTao || '';
		this.BacDaoTaoId = dotTuyenSinh.BacDaoTaoId || '';
		this.KhoaHoc = dotTuyenSinh.KhoaHoc || '';
		this.LoaiDangKyId = dotTuyenSinh.LoaiDangKyId || '';
		this.NamTuyenSinh = dotTuyenSinh.NamTuyenSinh || '';
		this.ThuTuHienThi = dotTuyenSinh.ThuTuHienThi || '';
		this.KichHoatDK = dotTuyenSinh.KichHoatDK || false;
		this.KichHoatXetTuyet = dotTuyenSinh.KichHoatXetTuyet || false;
		this.NgayTao = dotTuyenSinh.NgayTao || '';
		this.Ngaysua = dotTuyenSinh.Ngaysua || '';
		this.NguoiTao = dotTuyenSinh.NguoiTao || '';
		this.NguoiSua = dotTuyenSinh.NguoiSua || '';
		this.SoLuongNguyenVong = dotTuyenSinh.SoLuongNguyenVong || 0;
		this.HanDangKy = dotTuyenSinh.HanDangKy || '';
		this.NoiDungThongBao = dotTuyenSinh.NoiDungThongBao || '';
		this.NoiDungEmail = dotTuyenSinh.NoiDungEmail || '';
		this.DanhSachNganhDaoTao = dotTuyenSinh.DanhSachNganhDaoTao || new Array<NganhDaoTao>();
		this.CongThucKiemTraDieuKienNhapDiem = dotTuyenSinh.CongThucKiemTraDieuKienNhapDiem || '';
		this.ChoPhepGuiMail = dotTuyenSinh.ChoPhepGuiMail || false;
		this.ChoPhepNhapDCThuongTruVaTamTru = dotTuyenSinh.ChoPhepNhapDCThuongTruVaTamTru || false;
		this.KiemTraDieuKienNhapDiem = dotTuyenSinh.KiemTraDieuKienNhapDiem || false
		this.NgayBatDau = dotTuyenSinh.NgayBatDau || null;
		this.MaTienTo = dotTuyenSinh.MaTienTo || '';
		this.MaBacDaoTao = dotTuyenSinh.MaBacDaoTao || '';
		this.MaBatDau = dotTuyenSinh.MaBatDau || '';
		this.PhuongThucXetTuyenId = dotTuyenSinh.PhuongThucXetTuyenId || null;
		this.PhuongThucXetTuyens = dotTuyenSinh.PhuongThucXetTuyens || [];
		this.GiayToYeuCaus = dotTuyenSinh.GiayToYeuCaus || new Array<GiayTo>();
		this.ChoPhepUploadHoSo = dotTuyenSinh.ChoPhepUploadHoSo || false;
		this.FileHuongDan = dotTuyenSinh.FileHuongDan || new GiayTo();
		this.IsTaiFile = dotTuyenSinh.IsTaiFile || false;
		this.IsKhoa = dotTuyenSinh.IsKhoa || false;
		this.NgayKhoa = dotTuyenSinh.NgayKhoa || '';
		this.PhanQuyenChinhSuaHoSo = dotTuyenSinh.PhanQuyenChinhSuaHoSo || 0;
	}
}